chore: refactor all cli codebase using path alias
