+++ src/helpers/logNextSteps.ts
@@ -1,7 +1,7 @@
-import type { InstallerOptions } from "../installers/index.js";
-import { DEFAULT_APP_NAME } from "../consts.js";
-import { getUserPkgManager } from "../utils/getUserPkgManager.js";
-import { logger } from "../utils/logger.js";
+import type { InstallerOptions } from "~/installers/index.js";
+import { DEFAULT_APP_NAME } from "~/consts.js";
+import { getUserPkgManager } from "~/utils/getUserPkgManager.js";
+import { logger } from "~/utils/logger.js";
 
 // This logs the next steps that the user should take in order to advance the project
 export const logNextSteps = ({
