+++ src/index.ts
@@ -2,14 +2,14 @@
 import type { PackageJson } from "type-fest";
 import path from "path";
 import fs from "fs-extra";
-import { runCli } from "./cli/index.js";
-import { createProject } from "./helpers/createProject.js";
-import { initializeGit } from "./helpers/initGit.js";
-import { logNextSteps } from "./helpers/logNextSteps.js";
-import { buildPkgInstallerMap } from "./installers/index.js";
-import { logger } from "./utils/logger.js";
-import { parseNameAndPath } from "./utils/parseNameAndPath.js";
-import { renderTitle } from "./utils/renderTitle.js";
+import { runCli } from "~/cli/index.js";
+import { createProject } from "~/helpers/createProject.js";
+import { initializeGit } from "~/helpers/initGit.js";
+import { logNextSteps } from "~/helpers/logNextSteps.js";
+import { buildPkgInstallerMap } from "~/installers/index.js";
+import { logger } from "~/utils/logger.js";
+import { parseNameAndPath } from "~/utils/parseNameAndPath.js";
+import { renderTitle } from "~/utils/renderTitle.js";
 
 const main = async () => {
   renderTitle();
