+++ src/installers/prisma.ts
@@ -1,9 +1,9 @@
-import type { Installer } from "./index.js";
 import type { PackageJson } from "type-fest";
+import type { Installer } from "~/installers/index.js";
 import path from "path";
 import fs from "fs-extra";
-import { PKG_ROOT } from "../consts.js";
-import { execa } from "../utils/execAsync.js";
+import { PKG_ROOT } from "~/consts.js";
+import { execa } from "~/utils/execAsync.js";
 
 export const prismaInstaller: Installer = async ({
   projectDir,
