+++ src/utils/getT3Version.ts
@@ -1,7 +1,7 @@
 import type { PackageJson } from "type-fest";
 import path from "path";
 import fs from "fs-extra";
-import { PKG_ROOT } from "../consts.js";
+import { PKG_ROOT } from "~/consts.js";
 
 export const getVersion = () => {
   const packageJsonPath = path.join(PKG_ROOT, "package.json");
